removed duplication reverted ShaderDefVal move
