+++ crates/bevy_render/src/render_resource/pipeline.rs
@@ -1,17 +1,21 @@
 use crate::{
-    define_atomic_id,
-    render_resource::{BindGroupLayout, Shader, ShaderDefVal},
-    render_resource_wrapper,
+    define_atomic_id, render_resource::*, render_resource_wrapper, renderer::RenderDevice,
 };
 use bevy_asset::Handle;
 use std::{borrow::Cow, ops::Deref};
-use wgpu::{
-    BufferAddress, ColorTargetState, DepthStencilState, MultisampleState, PrimitiveState,
-    PushConstantRange, VertexAttribute, VertexFormat, VertexStepMode,
-};
 
 define_atomic_id!(RenderPipelineId);
 render_resource_wrapper!(ErasedRenderPipeline, wgpu::RenderPipeline);
+
+pub(crate) trait Pipeline<I, D, P> {
+    fn process_pipeline(
+        id: I,
+        descriptor: &D,
+        device: &RenderDevice,
+        shader_cache: &mut ShaderCache,
+        layout_cache: &mut LayoutCache,
+    ) -> PipelineState<P>;
+}
 
 /// A [`RenderPipeline`] represents a graphics pipeline and its stages (shaders), bindings and vertex buffers.
 ///
@@ -25,7 +29,7 @@ pub struct RenderPipeline {
 
 impl RenderPipeline {
     #[inline]
-    pub fn new(id: RenderPipelineId, value: wgpu::RenderPipeline) -> Self {
+    pub(crate) fn new(id: RenderPipelineId, value: wgpu::RenderPipeline) -> Self {
         Self {
             id,
             value: ErasedRenderPipeline::new(value),
@@ -47,6 +51,90 @@ impl Deref for RenderPipeline {
     }
 }
 
+impl Pipeline<RenderPipelineId, RenderPipelineDescriptor, RenderPipeline> for RenderPipeline {
+    fn process_pipeline(
+        id: RenderPipelineId,
+        descriptor: &RenderPipelineDescriptor,
+        device: &RenderDevice,
+        shader_cache: &mut ShaderCache,
+        layout_cache: &mut LayoutCache,
+    ) -> PipelineState<RenderPipeline> {
+        let vertex_module = match shader_cache.get(
+            device,
+            id.into(),
+            &descriptor.vertex.shader,
+            &descriptor.vertex.shader_defs,
+        ) {
+            Ok(module) => module,
+            Err(err) => {
+                return PipelineState::Err(err);
+            }
+        };
+
+        let fragment_data = if let Some(fragment) = &descriptor.fragment {
+            let fragment_module = match shader_cache.get(
+                device,
+                id.into(),
+                &fragment.shader,
+                &fragment.shader_defs,
+            ) {
+                Ok(module) => module,
+                Err(err) => {
+                    return PipelineState::Err(err);
+                }
+            };
+            Some((
+                fragment_module,
+                fragment.entry_point.deref(),
+                fragment.targets.as_slice(),
+            ))
+        } else {
+            None
+        };
+
+        let vertex_buffer_layouts = descriptor
+            .vertex
+            .buffers
+            .iter()
+            .map(|layout| RawVertexBufferLayout {
+                array_stride: layout.array_stride,
+                attributes: &layout.attributes,
+                step_mode: layout.step_mode,
+            })
+            .collect::<Vec<_>>();
+
+        let layout = descriptor
+            .layout
+            .as_ref()
+            .map(|layout| layout_cache.get(device, layout));
+
+        let descriptor = RawRenderPipelineDescriptor {
+            multiview: None,
+            depth_stencil: descriptor.depth_stencil.clone(),
+            label: descriptor.label.as_deref(),
+            layout,
+            multisample: descriptor.multisample,
+            primitive: descriptor.primitive,
+            vertex: RawVertexState {
+                buffers: &vertex_buffer_layouts,
+                entry_point: descriptor.vertex.entry_point.deref(),
+                module: &vertex_module,
+            },
+            fragment: fragment_data
+                .as_ref()
+                .map(|(module, entry_point, targets)| RawFragmentState {
+                    entry_point,
+                    module,
+                    targets,
+                }),
+        };
+
+        let pipeline = device.create_render_pipeline(id, &descriptor);
+
+        PipelineState::Ok(pipeline)
+    }
+}
+
 define_atomic_id!(ComputePipelineId);
 render_resource_wrapper!(ErasedComputePipeline, wgpu::ComputePipeline);
 
@@ -62,7 +150,7 @@ pub struct ComputePipeline {
 
 impl ComputePipeline {
     #[inline]
-    pub fn new(id: ComputePipelineId, value: wgpu::ComputePipeline) -> Self {
+    pub(crate) fn new(id: ComputePipelineId, value: wgpu::ComputePipeline) -> Self {
         Self {
             id,
             value: ErasedComputePipeline::new(value),
@@ -82,6 +170,44 @@ impl Deref for ComputePipeline {
     #[inline]
     fn deref(&self) -> &Self::Target {
         &self.value
+    }
+}
+
+impl Pipeline<ComputePipelineId, ComputePipelineDescriptor, ComputePipeline> for ComputePipeline {
+    fn process_pipeline(
+        id: ComputePipelineId,
+        descriptor: &ComputePipelineDescriptor,
+        device: &RenderDevice,
+        shader_cache: &mut ShaderCache,
+        layout_cache: &mut LayoutCache,
+    ) -> PipelineState<ComputePipeline> {
+        let compute_module = match shader_cache.get(
+            device,
+            id.into(),
+            &descriptor.shader,
+            &descriptor.shader_defs,
+        ) {
+            Ok(module) => module,
+            Err(err) => {
+                return PipelineState::Err(err);
+            }
+        };
+
+        let layout = descriptor
+            .layout
+            .as_ref()
+            .map(|layout| layout_cache.get(device, layout));
+
+        let descriptor = RawComputePipelineDescriptor {
+            label: descriptor.label.as_deref(),
+            layout,
+            module: &compute_module,
+            entry_point: descriptor.entry_point.as_ref(),
+        };
+
+        let pipeline = device.create_compute_pipeline(id, &descriptor);
+
+        PipelineState::Ok(pipeline)
     }
 }
