+++ crates/bevy_render/src/render_resource/shader.rs
@@ -1,4 +1,4 @@
-use crate::define_atomic_id;
+use crate::{define_atomic_id, render_resource::ShaderDefVal};
 use bevy_asset::{AssetLoader, AssetPath, Handle, LoadContext, LoadedAsset};
 use bevy_reflect::TypeUuid;
 use bevy_utils::{tracing::error, BoxedFuture, HashMap};
@@ -10,35 +10,6 @@ use thiserror::Error;
 use wgpu::{util::make_spirv, Features, ShaderModuleDescriptor, ShaderSource};
 
 define_atomic_id!(ShaderId);
-
-#[derive(Clone, PartialEq, Eq, Debug, Hash)]
-pub enum ShaderDefVal {
-    Bool(String, bool),
-    Int(String, i32),
-    UInt(String, u32),
-}
-
-impl From<&str> for ShaderDefVal {
-    fn from(key: &str) -> Self {
-        ShaderDefVal::Bool(key.to_string(), true)
-    }
-}
-
-impl From<String> for ShaderDefVal {
-    fn from(key: String) -> Self {
-        ShaderDefVal::Bool(key, true)
-    }
-}
-
-impl ShaderDefVal {
-    pub fn value_as_string(&self) -> String {
-        match self {
-            ShaderDefVal::Bool(_, def) => def.to_string(),
-            ShaderDefVal::Int(_, def) => def.to_string(),
-            ShaderDefVal::UInt(_, def) => def.to_string(),
-        }
-    }
-}
 
 #[derive(Error, Debug)]
 pub enum ShaderReflectError {
